haikuWhatsApp/backend/requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
airtable==0.4.8
annotated-types==0.6.0
anthropic==0.25.6
anyio==4.2.0
argcomplete==3.1.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
base58==2.1.1
bcrypt==4.0.1
borsh-construct==0.1.0
cachetools==4.2.4
certifi==2023.11.17
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.10
click==8.1.7
click-log==0.4.0
colorama==0.4.6
coloredlogs==15.0.1
construct==2.10.68
construct-typing==0.5.6
dataclasses-json==0.6.3
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.8
distro==1.9.0
docutils==0.20.1
dotty-dict==1.3.1
exceptiongroup==1.2.0
fastapi==0.99.1
filelock==3.12.4
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.9.2
gitdb==4.0.11
GitPython==3.1.41
gotrue==1.3.1
grpcio==1.62.2
h11==0.14.0
httpcore==0.16.3
httptools==0.6.0
httpx==0.23.3
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.0
install==1.3.5
invoke==1.7.3
jaraco.classes==3.3.0
jito_searcher_client==0.1.3
jsonalias==0.1.1
jsonpatch==1.33
jsonpointer==2.4
keyring==24.3.0
langchain==0.1.0
langchain-anthropic==0.1.11
langchain-community==0.0.34
langchain-core==0.1.46
langchain-openai==0.1.4
langchainplus-sdk==0.0.20
langsmith==0.1.51
marshmallow==3.20.2
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nh3==0.2.15
numexpr==2.8.7
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.16.0
openai==1.23.6
openapi-schema-pydantic==1.2.4
orjson==3.10.0
overrides==7.4.0
packaging==23.2
pandas==2.2.2
pipenv==2023.12.1
pipx==1.2.0
pkginfo==1.9.6
platformdirs==4.2.0
postgrest-py==0.10.3
posthog==3.0.2
protobuf==4.25.3
pulsar-client==3.3.0
pydantic==1.10.13
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
pypdf==3.16.4
PyPika==0.48.9
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-gitlab==3.15.0
python-multipart==0.0.6
python-semantic-release==7.33.0
pytz==2024.1
PyYAML==6.0.1
readme-renderer==42.0
realtime==1.0.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
semver==2.13.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
solana==0.30.1
solders==0.17.0
SQLAlchemy==2.0.25
starlette==0.27.0
storage3==0.5.0
sumtypes==0.1a6
supabase==1.0.0
supafunc==0.2.2
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.14.0
tomlkit==0.12.3
tqdm==4.66.1
tweepy==4.14.0
twilio==9.0.5
twine==3.8.0
typer==0.9.0
types-cachetools==4.2.10
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
userpath==1.8.0
uvicorn==0.23.2
uvloop==0.17.0
virtualenv==20.25.1
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.4
zipp==3.17.0
